Improve network check in certain situations.
I encountered a problem where the network was down but the DHCP/DNS router was responding to pings to "google.com" with its own router web page (hate that). This made the "ping google.com" test succeed and this program never kicked in. Fix this by pinging google's public DNS server 8.8.8.8 instead of google.com.

Also had tried an experimental test using Wget, but this did not succeed either, likely for the same reason. Left the code in there, but it's commented out.